vulkan: Comprehensive backend fixes and new shader support

+++ mlx/backend/vulkan/device.cpp
@@ -664,6 +664,7 @@ array make_scratch_view(const array& owner, Shape shape, Dtype dtype) {
 }
 
 struct StreamData {
+  std::thread::id owner_thread_id;
   vk::Semaphore timeline_semaphore;
   std::unique_ptr<SubmissionResources> recording_resources;
   std::vector<std::unique_ptr<SubmissionResources>> available_resources;
@@ -707,6 +708,14 @@ class VulkanDevice {
 
   void ensure_stream(int index) {
     (void)get_stream(index);
+  }
+
+  void validate_stream_thread(Stream s) {
+    auto* stream = get_stream(s.index);
+    if (stream->owner_thread_id != std::this_thread::get_id()) {
+      throw std::runtime_error(
+          "[vulkan] Stream accessed from a different thread than the one that created it.");
+    }
   }
 
   StreamData* get_stream(int index) {
@@ -826,6 +835,7 @@ class VulkanDevice {
   }
 
   void clear_streams() {
+    const auto current_thread = std::this_thread::get_id();
     vk::Device device;
     vk::Queue compute_queue;
     vk::Queue transfer_queue;
@@ -838,7 +848,13 @@ class VulkanDevice {
       has_transfer_queue = ctx.has_separate_transfer_queue();
     } catch (...) {
       std::lock_guard<std::mutex> lock(mutex_);
-      streams_.clear();
+      for (auto it = streams_.begin(); it != streams_.end();) {
+        if (it->second->owner_thread_id == current_thread) {
+          it = streams_.erase(it);
+        } else {
+          ++it;
+        }
+      }
       return;
     }
 
@@ -851,7 +867,12 @@ class VulkanDevice {
     }
 
     std::lock_guard<std::mutex> lock(mutex_);
-    for (auto& [_, stream] : streams_) {
+    for (auto it = streams_.begin(); it != streams_.end();) {
+      auto& stream = it->second;
+      if (stream->owner_thread_id != current_thread) {
+        ++it;
+        continue;
+      }
       destroy_submission_resources(device, stream->recording_resources);
       for (auto& resources : stream->available_resources) {
         destroy_submission_resources(device, resources);
@@ -865,8 +886,8 @@ class VulkanDevice {
         device.destroySemaphore(stream->timeline_semaphore);
         stream->timeline_semaphore = nullptr;
       }
+      it = streams_.erase(it);
     }
-    streams_.clear();
   }
 
   void synchronize_buffer_for_host_access(VulkanBuffer* buffer) {
@@ -2009,6 +2030,7 @@ class VulkanDevice {
   std::unique_ptr<StreamData> create_stream(int index) {
     auto stream = std::make_unique<StreamData>();
     stream->stream_index = index;
+    stream->owner_thread_id = std::this_thread::get_id();
 
     vk::SemaphoreTypeCreateInfo timeline_ci;
     timeline_ci.sType = vk::StructureType::eSemaphoreTypeCreateInfo;
@@ -2381,6 +2403,7 @@ void new_stream(Stream s) {
 }
 
 void synchronize(Stream s) {
+  mlx::core::vulkan::validate_stream_thread(s);
   mlx::core::vulkan::VulkanDevice::get().synchronize(s);
 }
 
@@ -2422,6 +2445,10 @@ void end_transfer_command_recording(int stream_index) {
 
 bool deferred_submission_active() {
   return deferred_submission_enabled();
+}
+
+void validate_stream_thread(Stream s) {
+  VulkanDevice::get().validate_stream_thread(s);
 }
 
 void retain_array_for_stream(const Stream& s, const array& arr) {
